Move aleph_client specific code to a module.
