+++ ink-wrapper-types/Cargo.toml
@@ -13,19 +13,27 @@ documentation = "https://docs.rs/ink-wrapper-types"
 keywords = ["ink", "wasm", "blockchain", "contract", "client"]
 
 [dependencies]
-aleph_client = { git = "https://github.com/Cardinal-Cryptography/aleph-node", package = "aleph_client", tag = "r-12.1", optional = true }
-
-anyhow = { version = "1.0.51", optional = true }
-async-trait = "0.1.64"
-
+# Base, common dependencies.
 ink_primitives = "4.3.0"
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = [
     "derive",
 ] }
+
+
+# aleph_client specific integration.
+aleph_client = { git = "https://github.com/Cardinal-Cryptography/aleph-node", package = "aleph_client", tag = "r-12.1", optional = true }
+anyhow = { version = "1.0.51", optional = true }
+async-trait = { version = "0.1.64", optional = true }
+subxt = { version = "0.30.1", optional = true }
 pallet-contracts-primitives = { git = "https://github.com/Cardinal-Cryptography/substrate.git", package = "pallet-contracts-primitives", branch = "aleph-v1.0.0", default-features = false, features = [
     "std",
-] }
-subxt = { version = "0.30.1", optional = true }
+], optional = true }
 
 [features]
-aleph_client = ["dep:aleph_client", "dep:anyhow", "dep:subxt"]
+aleph_client = [
+    "dep:aleph_client",
+    "dep:anyhow",
+    "dep:subxt",
+    "dep:async-trait",
+    "dep:pallet-contracts-primitives",
+]
